stage_5: remove test with undefined behavior
I had previously written some tests that change the value of a variable
and also reference the said variable in the same expression (e.g.
"(a = 4) / 4"). However, this leads to undefined behavior since the
order in which the operands of "/" get evaluated is implementation
defined, and the compiler is free to change that as it sees fit. The
only two binary operations where the order is established are && and ||.
So let's remove these tests.

More info: https://stackoverflow.com/a/27118744/11019779