stage_9: two extra tests for x86_64 calling conventions
One test checks that we correctly handle extra arguments that do not go
on registers, but the stack. The other checks that we do not overwrite
the argument saved in a register when generating the value for the
remaining arguments. We use division for this because it uses edx and
ecx, which respectively store the third and fourth arguments on a
function call.